+++ Dockerfile
@@ -1,4 +1,12 @@
-FROM node:20-alpine as builder
+FROM node:20-alpine as runner
+
+# hadolint ignore=DL3018
+RUN apk update && \
+  apk upgrade && \
+  apk add --update --no-cache tzdata && \
+  cp /usr/share/zoneinfo/Asia/Tokyo /etc/localtime && \
+  echo "Asia/Tokyo" > /etc/timezone && \
+  apk del tzdata
 
 WORKDIR /app
 
@@ -12,26 +20,10 @@ RUN echo network-timeout 600000 > .yarnrc && \
 COPY src src
 COPY tsconfig.json .
 
-RUN yarn package
-
-FROM node:20-alpine as runner
-
-# hadolint ignore=DL3018
-RUN apk update && \
-  apk upgrade && \
-  apk add --update --no-cache tzdata && \
-  cp /usr/share/zoneinfo/Asia/Tokyo /etc/localtime && \
-  echo "Asia/Tokyo" > /etc/timezone && \
-  apk del tzdata
-
-WORKDIR /app
-
-COPY --from=builder /app/output .
-
 ENV NODE_ENV production
 ENV API_PORT 80
 ENV GITHUB_USER_MAP_FILE_PATH /data/github-user-map.json
 
 VOLUME [ "/data" ]
 
-ENTRYPOINT [ "node", "index.js" ]
+ENTRYPOINT [ "yarn", "start" ]
