+++ package.json
@@ -7,12 +7,8 @@
   "main": "index.js",
   "repository": "git@github.com:book000/github-webhook-bridge.git",
   "scripts": {
-    "start": "ts-node -r tsconfig-paths/register ./src/main.ts",
-    "dev": "ts-node-dev --poll -r tsconfig-paths/register ./src/main.ts",
-    "package": "run-s clean compile packing",
-    "packing": "ncc build ./dist/main.js -o output/",
-    "compile": "tsc -p .",
-    "clean": "rimraf dist output",
+    "start": "tsx ./src/main.ts",
+    "dev": "tsx watch ./src/main.ts",
     "lint": "run-p -c lint:prettier lint:eslint lint:tsc",
     "lint:prettier": "prettier --check src",
     "lint:eslint": "eslint . --ext ts,tsx",
@@ -38,8 +34,7 @@
     "eslint-plugin-unicorn": "48.0.1",
     "fastify": "4.24.3",
     "prettier": "3.0.3",
-    "ts-node": "10.9.1",
-    "ts-node-dev": "2.0.0",
+    "tsx": "3.14.0",
     "typescript": "5.2.2",
     "yarn-run-all": "3.1.1"
   }
